sparse-index: integrate with reset
This change enables running the `reset` builtin command without expanding the
full index in a cone-mode sparse checkout.

`reset --soft` does not modify the index, so no compatibility changes are
needed for it to function without expanding the index. For all other reset
modes (`--mixed`, `--hard`, `--keep`, `--merge`), the full index is explicitly
expanded with `ensure_full_index` to maintain current behavior.

Additionally, the `read_cache()` check verifying an uncorrupted index is moved
after argument parsing and preparing the repo settings. The index is not used
by the preceding argument handling, but `read_cache()` does need to be run
after enabling sparse index for the command and before using the index for the
reset.

Signed-off-by: Victoria Dye <vdye@github.com>