reset: add test demonstrating mixed reset outside sparse checkout
Mixed reset in a sparse checkout (cone mode or otherwise) will, if a file
outside of the sparse definition has differences between the working tree
and target, write the file to the local working copy of the repository.
Since this behavior conflicts with the typical expectations of a strict
checkout definition/cone, a test is added demonstrating (and explaining) it.

Signed-off-by: Victoria Dye <vdye@github.com>