Revert "repo-settings: enable sparse index by default"
This reverts commit f55843a7b8701546cf35204d01bb5673638acfaf.

We are disabling the sparse index by default so we can take the latest
changes from the vfs-2.33.0 branch and cut a new release that should help
users affected by #450 (slow commands due to getting an environment
variable).

This commit represents an approach that would include shipping the latest FS
Monitor feature in addition to other installer fixes.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>